exepriment_backend_rest/requirements.txt
Django==4.0
djangorestframework==3.13.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2